mcp-server/requirements.txt
mcp>=0.1.0
click
docker>=7.0.0
python-dotenv
